+++ Include/Acidanthera/Library/OcConfigurationLib.h
@@ -358,6 +358,8 @@ typedef enum {
   _(OC_STRING                   , Comment          ,     , OC_STRING_CONSTR ("", _, __), OC_DESTR (OC_STRING) ) \
   _(BOOLEAN                     , Auxiliary        ,     , FALSE                       , ()                   ) \
   _(BOOLEAN                     , Enabled          ,     , FALSE                       , ()                   ) \
+  _(BOOLEAN                     , RealPath         ,     , FALSE                       , ()                   ) \
+  _(BOOLEAN                     , TextMode         ,     , FALSE                       , ()                   ) \
   _(OC_STRING                   , Name             ,     , OC_STRING_CONSTR ("", _, __), OC_DESTR (OC_STRING) ) \
   _(OC_STRING                   , Path             ,     , OC_STRING_CONSTR ("", _, __), OC_DESTR (OC_STRING) )
   OC_DECLARE (OC_MISC_TOOLS_ENTRY)