+++ Library/OcStorageLib/OcStorageLib.c
@@ -21,6 +21,7 @@
 #include <Library/DebugLib.h>
 #include <Library/DevicePathLib.h>
 #include <Library/MemoryAllocationLib.h>
+#include <Library/OcDevicePathLib.h>
 #include <Library/OcStringLib.h>
 #include <Library/OcStorageLib.h>
 #include <Library/UefiBootServicesTableLib.h>
@@ -215,8 +216,10 @@ EFI_STATUS
 OcStorageInitFromFs (
   OUT OC_STORAGE_CONTEXT               *Context,
   IN  EFI_SIMPLE_FILE_SYSTEM_PROTOCOL  *FileSystem,
-  IN  CONST CHAR16                     *Path,
-  IN  OC_RSA_PUBLIC_KEY                *StorageKey OPTIONAL
+  IN  EFI_HANDLE                       StorageHandle  OPTIONAL,
+  IN  EFI_DEVICE_PATH_PROTOCOL         *StoragePath   OPTIONAL,
+  IN  CONST CHAR16                     *StorageRoot,
+  IN  OC_RSA_PUBLIC_KEY                *StorageKey    OPTIONAL
   )
 {
   EFI_STATUS         Status;
@@ -228,7 +231,7 @@ OcStorageInitFromFs (
 
   ASSERT (Context != NULL);
   ASSERT (FileSystem != NULL);
-  ASSERT (Path != NULL);
+  ASSERT (StorageRoot != NULL);
 
   ZeroMem (Context, sizeof (*Context));
 
@@ -242,8 +245,8 @@ OcStorageInitFromFs (
 
   Status = SafeFileOpen (
     RootVolume,
-    &Context->StorageRoot,
-    (CHAR16 *) Path,
+    &Context->Storage,
+    (CHAR16 *) StorageRoot,
     EFI_FILE_MODE_READ,
     0
     );
@@ -251,7 +254,7 @@ OcStorageInitFromFs (
   RootVolume->Close (RootVolume);
 
   if (EFI_ERROR (Status)) {
-    DEBUG ((DEBUG_INFO, "OCST: Directory %s cannot be opened - %r\n", Path, Status));
+    DEBUG ((DEBUG_INFO, "OCST: Directory %s cannot be opened - %r\n", StorageRoot, Status));
     return Status;
   }
 
@@ -287,11 +290,14 @@ OcStorageInitFromFs (
   }
 
   gBS->InstallProtocolInterface (
-    &Context->StorageHandle,
+    &Context->DummyStorageHandle,
     &gEfiDevicePathProtocolGuid,
     EFI_NATIVE_INTERFACE,
     &mDummyBootDevicePath
     );
+  Context->StorageHandle   = StorageHandle;
+  Context->StoragePath     = StoragePath;
+  Context->StorageRoot     = StorageRoot;
   Context->DummyDevicePath = &mDummyBootDeviceFilePath.Vendor.Header;
   Context->DummyFilePath   = &mDummyBootDeviceFilePath.VendorFile.Header;
 
@@ -311,9 +317,9 @@ OcStorageFree (
   IN OUT OC_STORAGE_CONTEXT            *Context
   )
 {
-  if (Context->StorageRoot != NULL) {
-    Context->StorageRoot->Close (Context->StorageRoot);
-    Context->StorageRoot = NULL;
+  if (Context->Storage != NULL) {
+    Context->Storage->Close (Context->Storage);
+    Context->Storage = NULL;
   }
 
   if (Context->HasVault) {
@@ -345,12 +351,12 @@ OcStorageExistsFileUnicode (
     return TRUE;
   }
 
-  if (Context->StorageRoot == NULL) {
+  if (Context->Storage == NULL) {
     return FALSE;
   }
 
   Status = SafeFileOpen (
-    Context->StorageRoot,
+    Context->Storage,
     &File,
     (CHAR16 *) FilePath,
     EFI_FILE_MODE_READ,
@@ -393,7 +399,7 @@ OcStorageReadFileUnicode (
     return NULL;
   }
 
-  if (Context->StorageRoot == NULL) {
+  if (Context->Storage == NULL) {
     //
     // TODO: expand support for other contexts.
     //
@@ -401,7 +407,7 @@ OcStorageReadFileUnicode (
   }
 
   Status = SafeFileOpen (
-    Context->StorageRoot,
+    Context->Storage,
     &File,
     (CHAR16 *) FilePath,
     EFI_FILE_MODE_READ,
@@ -448,4 +454,81 @@ OcStorageReadFileUnicode (
   }
 
   return FileBuffer;
+}
+
+EFI_STATUS
+OcStorageGetInfo (
+  IN  OC_STORAGE_CONTEXT               *Context,
+  IN  CONST CHAR16                     *FilePath,
+  OUT EFI_DEVICE_PATH_PROTOCOL         **DevicePath   OPTIONAL,
+  OUT EFI_HANDLE                       *StorageHandle OPTIONAL,
+  OUT EFI_DEVICE_PATH_PROTOCOL         **StoragePath  OPTIONAL,
+  IN  BOOLEAN                          RealPath
+  )
+{
+  CHAR16   *FullPath;
+  UINTN    RootLength;
+  UINTN    FileSize;
+
+  if (RealPath
+    && Context->StorageHandle != NULL
+    && Context->StorageRoot != NULL
+    && Context->StoragePath != NULL) {
+
+    //
+    // Set the storage handle right away.
+    //
+    *StorageHandle = Context->StorageHandle;
+
+    //
+    // Prepare file path.
+    //
+    RootLength = StrLen (Context->StorageRoot);
+    FileSize   = StrSize (FilePath);
+    FullPath   = AllocatePool (RootLength * sizeof (CHAR16) + sizeof (CHAR16) + FileSize);
+
+    if (FullPath == NULL) {
+      return EFI_OUT_OF_RESOURCES;
+    }
+
+    CopyMem (&FullPath[0], Context->StorageRoot, RootLength * sizeof (CHAR16));
+    FullPath[RootLength] = '\\';
+    CopyMem (&FullPath[RootLength + 1], FilePath, FileSize);
+
+    //
+    // Set joint device path.
+    //
+    *DevicePath = AppendFileNameDevicePath (Context->StoragePath, FullPath);
+    if (*DevicePath == NULL) {
+      FreePool (FullPath);
+      return EFI_OUT_OF_RESOURCES;
+    }
+
+    //
+    // Set onstorage path.
+    //
+    *StoragePath = FileDevicePath (NULL, FullPath);
+    if (*StoragePath == NULL) {
+      FreePool (*DevicePath);
+      FreePool (FullPath);
+      return EFI_OUT_OF_RESOURCES;
+    }
+
+    return EFI_SUCCESS;
+  }
+
+  *StorageHandle = Context->DummyStorageHandle;
+
+  *DevicePath = DuplicateDevicePath (Context->DummyDevicePath);
+  if (*DevicePath == NULL) {
+    return EFI_OUT_OF_RESOURCES;
+  }
+
+  *StoragePath = DuplicateDevicePath (Context->DummyFilePath);
+  if (*StoragePath == NULL) {
+    FreePool (*DevicePath);
+    return EFI_OUT_OF_RESOURCES;
+  }
+
+  return EFI_SUCCESS;
 }