+++ Changelog.md
@@ -6,6 +6,8 @@ OpenCore Changelog
 - Dropped HII services from OpenDuet improving size and performance
 - Fixed patching of injected kexts in mkext
 - Added support for launching from relative paths
+- Added direct path passing for tools via `RealPath`
+- Allowed launching tools and entries in text mode via `TextMode`
 
 #### v0.6.3
 - Added support for xml comments in plist files